Wellness Chatbot/requirements.txt
flask
flask-cors
openai
authlib
python-dotenv
requests
pymongo
